+++ test/expected/default_options.js
@@ -2,7 +2,7 @@ angular.module('module1', [])
 
 .constant('constant1', {global_key:'global_value',key1:'value1',key2:'value2'})
 
-.constant('constant2', undefined)
+.constant('constant2', null)
 
 .value('value1', {key1:'value1'})